1. 알뜰구매용 블록 추가

+++ server/src/main/java/com/anamensis/server/mapper/MemberMapper.java
@@ -30,7 +30,7 @@ public interface MemberMapper {
 
     int updatePoint(
             @Param("id") long id,
-            @Param("point") int point
+            @Param("point") long point
     );
 
     int updatePwd (
